moved load_genome_DB to databases, some restructuring in classify_genome
